+++ tests/AbstractKeyValueStoreTest.php
@@ -262,6 +262,27 @@ abstract class AbstractKeyValueStoreTest extends PHPUnit_Framework_TestCase
         $this->store->get('foo');
     }
 
+    public function testGetIfExists()
+    {
+        $this->store->set('foo', 'bar');
+
+        $this->assertSame('bar', $this->store->getIfExists('foo', 'baz'));
+    }
+
+    /**
+     * @dataProvider provideInvalidKeys
+     * @expectedException \Webmozart\KeyValueStore\Api\InvalidKeyException
+     */
+    public function testGetIfExistsFailsIfInvalidKey($key)
+    {
+        $this->store->getIfExists($key);
+    }
+
+    public function testGetIfExistsReturnsDefaultIfKeyNotFound()
+    {
+        $this->assertSame('bar', $this->store->getIfExists('foo', 'bar'));
+    }
+
     public function testGetMultipleIgnoresArrayIndices()
     {
         $this->store->set('a', 1234);